Moved my main file here
There are some major problems IDK how to fix. GG go next. This was a cool learning experience and a nice way to dabble my feet with this language. Idk if i will ever come back, but if i do i will create a ui for this and maybe multiplayer.