Great update
So I left school bc i feel nausea and want to throwup but finisihing my code is more important. Anyways It should be fully functional and some rules might  not be adapted in my version of Black jack but its okay. Maybe making ui for this is my next step. Im really glad this basically worked fine the frst time i ran this.