Kaguya Sama Season 3 ep 1 is out
I had alot of ideas in my head for making a better bot but the more i think about it, the more i get confused. Honestly rn all i know is that my syntax is good but i dont know how this really works anymore. The logic is getting harder to connect the more i add to this program.